requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pydantic-settings
pymysql